1. Whoops, side listeners have to use longs to track session stats.
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/alpine@3293 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ mvvm/main/com/metavize/mvvm/argon/SideListener.java
@@ -18,11 +18,11 @@ import com.metavize.jvector.SourceEndpointListener;
 
 class SideListener implements SinkEndpointListener, SourceEndpointListener
 {
-    protected int txChunks = 0;
-    protected int txBytes  = 0;
+    protected long txChunks = 0;
+    protected long txBytes  = 0;
 
-    protected int rxChunks = 0;
-    protected int rxBytes  = 0;
+    protected long rxChunks = 0;
+    protected long rxBytes  = 0;
 
     protected boolean isSourceShutdown = false;
     protected boolean isSinkShutdown   = false;
